tables: fix look up of special characters in Code::from_map
There were two issues: one with the control map getting iterated once
for each iteration of the code map, the second with the body of the
control map check, which was actually comparing code_map with the passed
in character, not control_map.

This meant special characters never got looked up.